LPS-131312 Update and rename classes
