+++ modules/apps/portal/portal-cluster-multiple-sample-web/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Portal Cluster Multiple Sample Web
+Bundle-SymbolicName: com.liferay.portal.cluster.multiple.sample.web
+Bundle-Version: 1.0.0
