+++ modules/apps/portal/portal-cluster-multiple-sample-web/src/main/java/com/liferay/portal/cluster/multiple/sample/web/internal/ClusterSampleData.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cluster.multiple.sample.internal;
+package com.liferay.portal.cluster.multiple.sample.web.internal;
 
 import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.StringUtil;
