+++ modules/apps/portal/portal-cluster-multiple-sample-web/src/main/java/com/liferay/portal/cluster/multiple/sample/web/internal/ClusterSampleDispatcher.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.cluster.multiple.sample.internal;
+package com.liferay.portal.cluster.multiple.sample.web.internal;
 
-import com.liferay.portal.cluster.multiple.sample.configuration.ClusterSampleConfiguration;
+import com.liferay.portal.cluster.multiple.sample.web.configuration.ClusterSampleConfiguration;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.cluster.ClusterExecutor;
 import com.liferay.portal.kernel.cluster.ClusterMasterExecutor;
@@ -42,7 +42,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Tina Tian
  */
 @Component(
-	configurationPid = "com.liferay.portal.cluster.multiple.sample.configuration.ClusterSampleConfiguration",
+	configurationPid = "com.liferay.portal.cluster.multiple.sample.web.configuration.ClusterSampleConfiguration",
 	immediate = true, service = {}
 )
 public class ClusterSampleDispatcher {
