+++ modules/apps/portal/portal-cluster-multiple-sample-web/src/main/resources/META-INF/resources/css/main.scss
@@ -1,0 +1,2 @@
+.portal-cluster-multiple-sample-web {
+}
