--- modules/apps/portal/portal-cluster-multiple-sample/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Portal Cluster Multiple Sample
-Bundle-SymbolicName: com.liferay.portal.cluster.multiple.sample
-Bundle-Version: 1.0.0
