--- modules/apps/portal/portal-cluster-multiple-sample/src/main/resources/META-INF/resources/css/main.scss
@@ -1,2 +1,0 @@
-.portal-cluster-multiple-sample {
-}
